+++ crates/re_renderer/examples/renderer_standalone.rs
@@ -18,6 +18,8 @@ use instant::Instant;
 use itertools::Itertools;
 use macaw::IsoTransform;
 use rand::Rng;
+use smallvec::smallvec;
+
 use re_renderer::{
     config::{supported_backends, HardwareTier, RenderContextConfig},
     renderer::{
@@ -237,14 +239,14 @@ fn build_lines(re_ctx: &mut RenderContext, seconds_since_startup: f32) -> LineDr
         &[
             // Complex orange line.
             LineStrip {
-                points: lorenz_points,
+                points: lorenz_points.into(),
                 radius: 0.05,
                 color: [255, 191, 0, 255],
                 flags: LineStripFlags::empty(),
             },
             // Green Zig-Zag
             LineStrip {
-                points: vec![
+                points: smallvec![
                     glam::vec3(0.0, -1.0, 0.0),
                     glam::vec3(1.0, 0.0, 0.0),
                     glam::vec3(2.0, -1.0, 0.0),
@@ -555,9 +557,7 @@ fn main() {
 
     #[cfg(not(target_arch = "wasm32"))]
     {
-        if std::env::var("RUST_LOG").is_err() {
-            std::env::set_var("RUST_LOG", "info,wgpu_core=warn,wgpu_hal=warn");
-        }
+        re_log::set_default_rust_log_env();
         tracing_subscriber::fmt::init();
 
         // Set size to a common physical resolution as a comparable start-up default.
@@ -597,7 +597,7 @@ fn redirect_tracing_to_console_log() {
     tracing::subscriber::set_global_default(
         tracing_subscriber::Registry::default()
             .with(tracing_subscriber::EnvFilter::new(
-                "debug,wgpu_core=warn,wgpu_hal=warn",
+                re_log::default_log_filter(),
             ))
             .with(tracing_wasm::WASMLayer::new(
                 tracing_wasm::WASMLayerConfig::default(),